% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyArea.R
\name{area_polygon}
\alias{area_polygon}
\alias{area_polygon.default}
\alias{area_polygon.formula}
\alias{print.area_polygon}
\title{Polygon area}
\usage{
area_polygon(x, ...)

\method{area_polygon}{default}(x, y = NULL, ...)

\method{area_polygon}{formula}(x, data = parent.frame(), ...)

\method{print}{area_polygon}(x, ...)
}
\arguments{
\item{x, y}{numeric vectors defining vertices in order.  Any method
recognized by grDevices::xy.coords() is accepted.
Or a formula for other methods.  Or an object of class "area_polygon"}

\item{...}{Additional arguments, currently not used}

\item{data}{a data frame}
}
\value{
An object of class "area_polygon". A single numerical value with the polygon area
    with an attribute giving the sign, as determined by the orientation of the vertices.
}
\description{
Calculate the area of a non-intersecting polygon given its vertices
}
\examples{
z <- complex(real = rnorm(200), imaginary = rnorm(200))
hz <- z[chull(z)]
plot(hz)
polygon(hz)
points(z, pch = ".", cex = 2, col = "red")
text(mean(hz), labels=paste("area =", area_polygon(hz)))
}
