% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{arrows}
\alias{arrows}
\alias{arrows,numeric,numeric-method}
\alias{arrows,xy,xy-method}
\alias{arrows,xy,missing-method}
\title{Arrows}
\usage{
arrows(x0, y0, ...)

\S4method{arrows}{numeric,numeric}(x0, y0, x1, y1, length = 4, angle = 15, gap, ...)

\S4method{arrows}{xy,xy}(x0, y0, ...)

\S4method{arrows}{xy,missing}(x0, y0, ..., circular = FALSE)
}
\arguments{
\item{x0, y0}{two numeric vectors of a single object of class \code{"xy"}}

\item{...}{additional arguments passed on to \code{graphics::segments}}

\item{x1, y1}{two numeric vectors of a single object of class \code{"xy"}}

\item{length}{length of the arrow head barb, in MILLIMETRES, see \code{\link{in2mm}}}

\item{angle}{as for \code{graphics::arrows}}

\item{gap}{numeric of length 1 or 2 giving the size of any gap to be left
between the arrow and the points which it connects, in MILLIMETRES}

\item{circular}{logical: should the arrows link up with the initial point? (Single location argument only.)}
}
\value{
invisible null value
}
\description{
Front end to the \code{graphics::arrows} function allowing
simplified specification of end points
}
\examples{
z <- with(roundTrip, setNames(complex(real = Longitude, imaginary = Latitude), Locality))
plot(z, asp = 1, pch = 20, cex = 0.7, xlab = "Longitude", ylab = "Latitude")
arrows(z, cyc(z), col = "red", gap = c(0,1.5))
}
