% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{as.function.density}
\alias{as.function.density}
\title{Density to function}
\usage{
\method{as.function}{density}(x, ...)
}
\arguments{
\item{x}{A kernel density object as produced by \code{\link{density}}}

\item{...}{ignored}
}
\value{
An R function of one argument providing the KDE at any specific set of values
}
\description{
Coercion function for kernel density object into R functions
}
