% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{barplot}
\alias{barplot}
\title{Barplot}
\usage{
barplot(
  height,
  ...,
  border = colour,
  colour = color,
  color = NULL,
  col = fill,
  fill = NULL
)
}
\arguments{
\item{height}{As for \code{graphics::barplot}}

\item{...}{additional arguments sent to \code{graphics::barplot}}

\item{border, colour, color}{alternative argument names for the border; may be a colour palette}

\item{col, fill}{alternative argument names for the fill colour; may be a colour palette}
}
\value{
as for \code{graphics::barplot}
}
\description{
Interface to \code{graphics::barplot} with extra flexibility
}
\examples{
## Poor man's flashy histogram
rnorm(5000) \%>\% cut(c(-Inf,(-5:5)/2,Inf)) \%>\%
   table \%>\% barplot(fill=pal_cool, space=0, las=2)
}
