% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{binNumber}
\alias{binNumber}
\title{Bin numbers for histograms}
\usage{
binNumber(n, ...)
}
\arguments{
\item{n}{The required precise number of bins to be used, unaligned.}

\item{...}{extra arguments, currently unused}
}
\value{
a breaks function to be used by \code{hist}
}
\description{
Generates a function to be used as the \code{breaks=} argument of
\code{hist}, to ensure that a fixed number of bins will be used, precisely.
In this case specifying a nice alignment is not available.
}
\examples{
hist(~medv/rm, MASS::Boston, binNumber(25), fill = pal_green2brown)
with(MASS::Boston, rug((medv/rm), col = "red"))
hist(~log10(medv/rm), MASS::Boston, binNumber(25), fill = pal_green2brown)
with(MASS::Boston, rug(log10(medv/rm), col = "red"))
}
