% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{binWidth}
\alias{binWidth}
\title{Bin Widths for Histograms}
\usage{
binWidth(w, align = NULL, ...)
}
\arguments{
\item{w}{the required bin width in x-units}

\item{align}{an optional value at which some bin boundary will align.}

\item{...}{additional parameters, currently not used}
}
\value{
a breaks function to be used by \code{hist}
}
\description{
Generates a function to be used as the \code{breaks=} argument of
\code{hist}, to ensure bins of a given, fixed width will be used
}
\examples{
hist(~log(crim), MASS::Boston, binWidth(0.35, align = 0), fill = pal_desert)
}
