% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{bivariateMedian}
\alias{bivariateMedian}
\alias{coordinateMedian}
\alias{spatialMedian}
\alias{ojaMedian}
\title{Bivariate medians}
\usage{
coordinateMedian(x, y = NULL, na.rm = FALSE)

spatialMedian(x, y = NULL, na.rm = FALSE)

ojaMedian(x, y = NULL, na.rm = FALSE)
}
\arguments{
\item{x, y}{any of the known ways to specify bivariate points in traditional graphics}

\item{na.rm}{logical: should any missing values be removed prior to the computation?}
}
\value{
a complex number giving the bivariate median as its real and complex parts
}
\description{
Bivariate medians
}
