% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{boxplot}
\alias{boxplot}
\alias{boxplot,ANY,missing,missing,missing-method}
\alias{boxplot,ANY,ANY,ANY,ANY-method}
\title{Boxplot}
\usage{
boxplot(x, ..., fill, color, colour)

\S4method{boxplot}{ANY,missing,missing,missing}(x, ..., fill, color, colour)

\S4method{boxplot}{ANY,ANY,ANY,ANY}(x, ..., fill, color, colour)
}
\arguments{
\item{x}{as for \code{graphics::boxplot}}

\item{...}{additional arguments passed on to \code{graphics::boxplot}}

\item{fill}{alternative argument name to \code{col} for the fill colour; may be a colour palette}

\item{color, colour}{alternative argument names for \code{border} for the colour of the borders; may be a colour palette}
}
\value{
as for \code{graphics::boxplot}
}
\description{
Alternative interface to \code{graphics::boxplot}
}
\examples{
## Poor man's faceting
boxplot(Days ~ Age+Sex, MASS::quine, fill = pal_green2brown(4),
    axes = FALSE, ann = FALSE)
axis(2)
box()
with(MASS::quine, legend("topleft", levels(Age),
     fill = pal_green2brown(4), title = "Age group", bty = "n"))
pu <- par("usr")[1:2]
M <- matrix(c(3, 1, 1, 3)/4, ncol = 2)
abline(v = mean(pu), lwd = 2)
mtext(c("Female", "Male"), line = 0.5, at = M \%*\% pu)
}
