% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{chull}
\alias{chull}
\title{Convex Hull}
\usage{
chull(x, y = NULL, peel = 0)
}
\arguments{
\item{x, y}{Numeric vectors of equal length or any specification of a set of
points as allowed by grDevices::xy.coords()}

\item{peel}{The number of convex hulls to peel off the set before taking the
convex hull of the remaining points}
}
\value{
An index vector giving the points in the original set forming the
    convex hull after peeling
}
\description{
Find the convex hull of a planar set of points allowing for peeling.
}
\examples{
z <- complex(real = rnorm(50), imaginary = rnorm(50))
plot(z)
polygon(z[chull(z)])
polygon(z[chull(z, peel = 1)], border = "red")
polygon(z[chull(z, peel = 2)], border = "blue")
}
