% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playingWithColours.R
\name{closest_colours}
\alias{closest_colours}
\alias{dist_rgb}
\alias{dist_lab}
\title{Find Closest Colours}
\usage{
closest_colours(colours, reference = colors())

dist_rgb(colours)

dist_lab(colours)
}
\arguments{
\item{colours}{The search colours, in any form but usually in hex}

\item{reference}{The reference set of colours}
}
\value{
A set of colours in the reference set closest to those in the search set;
        for \code{colour_distance} a \code{dist} object of pairwise distances.
}
\description{
Find the colours in a reference set closest to a specified set of
colours (in RGB space with L1 norm).
}
\examples{
set.seed(1234)
rc <- random_colours(10)
cols <- as.vector(rbind(rc, closest_colours(rc)))
showColours(cols)
d <- dist_lab(cols)
ds <- cmdscale(d)/1000
plot(ds, asp = 1, type = "n", bty = "n",
     xlab = "First coordinate", ylab = "Second coordinate")
text(ds, labels = cols, pos = avoid(ds), cex = 0.7,
     offset = 0.35, xpd = NA)
i <- 2*(1:10)
segments(ds[i-1, ], ds[i, ])
points(ds, pch = 19, col = cols)
}
