% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playingWithColours.R
\name{col2hex}
\alias{col2hex}
\title{Colours to Hex Format}
\usage{
col2hex(colours, names_ = nam)
}
\arguments{
\item{colours}{The character string vector of colours}

\item{names_}{The names to be given to the components of the result}
}
\value{
A character string vector of the colours in hex format
}
\description{
Convert a vector of colours in any form to hex
}
\examples{
col2hex(colors()[5:15])
}
