% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cq}
\alias{cq}
\title{Unquited character concatenation}
\usage{
cq(...)
}
\arguments{
\item{...}{names, normally unquoted, used as strings to be concatenated.  Strings must be quoted if non-standard R names.}
}
\value{
A character string vector
}
\description{
Allows character strings to be concatenated without quotes, unless
the string itself is a non-standard R name
}
\examples{
vars <- cq(alpha, beta, gamma, "_delta_")
}
