% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cyc}
\alias{cyc}
\title{Circular cycle function}
\usage{
cyc(z, k = 1)
}
\arguments{
\item{z}{Any object with indexed components}

\item{k}{integer. The number of steps to cycle: positive: first-to-last, negative: last-to-first}
}
\value{
a structure similar to \code{z} with its components re-ordered by cycling
}
\description{
Permute the components of a vector or vector-like object in a circular pattern
}
\examples{
cyc(letters, -3)
}
