% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{density_histogram}
\alias{density_histogram}
\alias{print.density_histogram}
\title{Density histogram}
\usage{
density_histogram(x, ..., polygon = TRUE, plot)

\method{print}{density_histogram}(
  x,
  ...,
  xlab = deparse(x$call$x),
  ylab = "Density",
  fill = alpha("sky blue", 0.5),
  colour = alpha("steel blue", 0.5)
)
}
\arguments{
\item{x, ...}{As for \code{hist}, or an object of class \code{"density_histogram"} for the \code{print} method.}

\item{polygon}{logical: suitable for use with \code{polygon()} rather than \code{lines()}?}

\item{plot}{Ignored. Plotting is always suppressed}
}
\value{
A list with components including \code{x} and \code{y} suitable for superimposing on other graphics with \code{lines()}
}
\description{
Construct a density histogram (i.e. \code{probability = TRUE}) as an xy-list
in a form suitable for superimposing on other graphics, e.g. density plots.
NOTE: this is NOT a plotting function.  Use \code{hist(x, probability = TRUE)}
for direct plots.
}
\examples{
d <- density(~log(crim), MASS::Boston)
h <- density_histogram(~log(crim), MASS::Boston, binWidth(1/3))
plot(d, xlim = xrange(d, h), ylim = yrange(d, h),
     main = "", xlab = "log(crim)", col = "blue")
lines(update(h, polygon = FALSE), col = alpha("red", 0.5))
polygon(h, fill = alpha("red", 0.5),
        colour = "transparent")
}
