% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DensityOps}
\alias{DensityOps}
\alias{Ops.density}
\title{Arithmetic operations on density estimates}
\usage{
\method{Ops}{density}(e1, e2)
}
\arguments{
\item{e1}{either \code{"numeric"} or \code{"density"} object}

\item{e2}{either \code{"numeric"} or \code{"density"} object}
}
\value{
A KDE, or KDE-like object representing the operation.
}
\description{
This facility is mainly intended for generating mixtures of KDEs
}
\examples{
set.seed(1234)
y1 <- rnorm(90000, 0.0, 1.0)
y2 <- rnorm(10000, 2.5, 0.5)
d1 <- density(y1)
d2 <- density(y2)
d12 <- 0.9*d1 + 0.1*d2  ## 5\% contaminated normal
plot(d12, main = "Normal with 5\% contamination")
grid(lty = "dashed")
lines(0.9*d1, col = "dark green", lwd = 1.5)
lines(0.1*d2, col = "red", lwd = 1.5)
curve(0.9*dnorm(x, 0, 1) + 0.1*dnorm(x, 2.5, 0.5),
      col = "purple", lty = "dashed", add = TRUE)
dyy <- density(c(y1, y2))
plot(dyy, type = "n",
     main = "Density of mixture -vs- mixture of densities")
grid(lty = "dashed")
polygon(density_histogram(c(y1, y2), binWidth(0.3)),
        fill = alpha("sky blue", 0.5),
        colour = "transparent")
lines(dyy, col = "dark green")
lines(d12, col = "red", lty = "dashed", lwd = 2)
}
