% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{getColors}
\alias{getColors}
\alias{getColours}
\title{Extended colour sets}
\usage{
getColors(
  description,
  database = c("Set1", "Set2", "set1", "set2", "1", "2"),
  ignore.case = TRUE
)

getColours(
  description,
  database = c("Set1", "Set2", "set1", "set2", "1", "2"),
  ignore.case = TRUE
)
}
\arguments{
\item{description}{A regular expression fot the colour name}

\item{database}{Either \code{"Set1"} (the default) or \code{"Set2"}.}

\item{ignore.case}{Logical.  Should case be ignored in the \code{description}?}
}
\value{
A vector of colours matching the \code{description}, or \code{character(0)}.
}
\description{
Search an estended database for more colours using a regular
expression on their given names
}
\examples{
getColors("french")  ## Any colour names with "French" as a component.
}
