% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hist}
\alias{hist}
\alias{hist,numeric-method}
\alias{hist,formula-method}
\title{Histograms}
\usage{
hist(x, ...)

\S4method{hist}{numeric}(
  x,
  ...,
  border = colour,
  colour = color,
  color = NULL,
  col = fill,
  fill = NULL,
  main,
  xlab
)

\S4method{hist}{formula}(x, data = parent.frame(2), ..., main, xlab)
}
\arguments{
\item{x}{either a numerical vector or a one-sided formula specifying the response variable}

\item{...}{additional arguments sent to \code{graphics::hist}}

\item{border, colour, color}{alternative argument names for the border}

\item{col, fill}{alternative argument names for the fill colour; may be a colout palette function}

\item{main, xlab}{as for \code{graphics::hist}}

\item{data}{a data frame or environment}
}
\value{
as for \code{graphics::hist}
}
\description{
Front-end to \code{\link[graphics]{hist}} allowing a formula interface alternative
}
\examples{
hist(~log10(crim), MASS::Boston, binWidth(0.125, align = 0),
     fill = pal_sea2sand, colour = "grey", lwd = 0.5, prob = TRUE)
rug(~log10(crim), MASS::Boston, col = "hot pink")
with(MASS::Boston, lines(density(log10(crim)), col="brown"))
}
