% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{is.colour}
\alias{is.colour}
\alias{is.color}
\title{Check colour string}
\usage{
is.colour(x)

is.color(x)
}
\arguments{
\item{x}{A character string vector}
}
\value{
A logical vector indicating validity
}
\description{
Check if character strings can be used as colour specifications
}
\examples{
is.colour(c("dark grey", "horrible yellow", 1))
is.colour(1)
}
