% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{myPalette}
\alias{myPalette}
\title{Generate a colour palette}
\usage{
myPalette(...)
}
\arguments{
\item{...}{colour names for the palette, possibly unquoted}
}
\value{
a colour palette function
}
\description{
Generate a colour palette
}
\examples{
pal <- myPalette(pink, beige, 'sky blue')
showColours(pal, n = 50)
}
