% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{palettes}
\alias{palettes}
\alias{pal_cool}
\alias{pal_heat_hcl}
\alias{pal_cool_hcl}
\alias{pal_pastels}
\alias{pal_blue2red}
\alias{pal_sea2sand}
\alias{pal_desert}
\alias{pal_green2brown}
\alias{pal_green2pink}
\alias{pal_grey}
\alias{pal_gray}
\alias{pal_namatjira}
\alias{pal_dark_rainbow}
\title{Colour palette functions}
\usage{
pal_cool(n)

pal_heat_hcl(n)

pal_cool_hcl(n)

pal_pastels(n)

pal_blue2red(n)

pal_sea2sand(n)

pal_desert(n)

pal_green2brown(n)

pal_green2pink(n)

pal_grey(n)

pal_gray(n)

pal_namatjira(n)

pal_dark_rainbow(n)
}
\arguments{
\item{n}{desired number of colours}
}
\value{
character string vector of colour strings
}
\description{
Colour palette functions
}
\examples{
barplot(1:15, fill = pal_sea2sand, colour = "red", axes = FALSE) -> bp
z <- complex(real = bp,imaginary = seq_along(bp))
points(z, xpd = NA)
arrows(z[-length(z)], z[-1], gap = 1.5, col = "black")
labels <- c("sea", gsub(" ", "0", format(2:14)), "sand")
text(z, labels, pos = 3, xpd = NA, col = pal_blue2red(15))
}
