% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pie}
\alias{pie}
\title{Pie charts}
\usage{
pie(
  x,
  ...,
  border = colour,
  colour = color,
  color = NULL,
  col = fill,
  fill = NULL
)
}
\arguments{
\item{x}{as for \code{graphics::pie}}

\item{...}{additional arguments sent to \code{graphics::pie}}

\item{border, colour, color}{alternative argument names for the border; may be a colour palette}

\item{col, fill}{alternative argument names for the fill colour; may be a colour palette}
}
\value{
as for \code{graphics::pie}
}
\description{
Alternative interface to \code{graphics::pie} with extra flexibility
}
\examples{
## A random number of pie slices
pie(rep(1, rpois(1, 15)), fill = pal_blue2red, colour = "grey")
}
