% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{polygon}
\alias{polygon}
\title{Polygon}
\usage{
polygon(
  x,
  ...,
  border = colour,
  colour = color,
  color = NULL,
  col = fill,
  fill = NULL
)
}
\arguments{
\item{x}{as for \code{graphics::polygon}}

\item{...}{additional arguments sent to \code{graphics::polygon}}

\item{border, colour, color}{alternative argument names for the border}

\item{col, fill}{alternative argument names for the fill colour}
}
\value{
as for \code{graphics::polygon}
}
\description{
An alternative interface to the \code{graphics::polygon} function
}
\examples{
set.seed(1234)
z <- complex(,runif(50), rnorm(50))
oz <- ojaMedian(z)
z <- z[order(Arg(z - oz))]
plot(z, type = "n", axes = FALSE, ann = FALSE)
polygon(z, fill = "sandy brown", colour = "grey")
points(oz, pch=3, cex = 2)
segments(oz, z, lwd = 0.5)
}
