% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playingWithColours.R
\name{random_colours}
\alias{random_colours}
\title{Random Colours}
\usage{
random_colours(n)
}
\arguments{
\item{n}{integer: the number of colours required}
}
\value{
A character string vector of the colours in hex format
}
\description{
Generate a set of random colours for demonstration purposes
}
\examples{
set.seed(1234)
showColours(random_colours(20))
}
