% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{rglColors}
\alias{rglColors}
\title{Surface colours for rgl}
\usage{
rglColors(
  z,
  pal = pal_pastels,
  no = length(unique(z)),
  col = fill,
  fill = pal(no)
)
}
\arguments{
\item{z}{A numeric matrix of heights}

\item{pal}{a colour palette to use}

\item{no}{the number of colours to use}

\item{col, fill}{Alternative names for the colours to use}
}
\value{
a colour vector suitable for use with rgl surface displays
}
\description{
Generate colours for use in a surface display with package rgl
}
\examples{
cols <- rglColors(volcano)
}
