% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rug}
\alias{rug}
\alias{rug,formula-method}
\title{Rug}
\usage{
rug(x, ...)

\S4method{rug}{formula}(x, data = sys.parent(2), ...)
}
\arguments{
\item{x}{A numeric vector or one-sided formula}

\item{...}{additional arguments for \code{\link[graphics]{rug}}}

\item{data}{A data frame or environment used if \code{x} is a formula}
}
\value{
As for \code{\link[graphics]{rug}}.
}
\description{
Front-end to \code{\link[graphics]{rug}} allowing a formula interface
}
\examples{
hist(~medv/rm, MASS::Boston)
rug(~medv/rm, MASS::Boston, col = "red")
}
