% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{segments}
\alias{segments}
\alias{segments,xy,xy-method}
\alias{segments,xy,missing-method}
\title{Segments}
\usage{
segments(x0, y0, ...)

\S4method{segments}{xy,xy}(x0, y0, ...)

\S4method{segments}{xy,missing}(x0, y0, ..., circular = FALSE)
}
\arguments{
\item{x0, y0}{two numeric vectors of a single object of class \code{"xy"}}

\item{...}{additional arguments passed on to \code{graphics::segments}}

\item{circular}{logical: should the line link up with the initial point?  (Single location argument only.)}

\item{x1, y1}{two numeric vectors of a single object of class \code{"xy"}}
}
\value{
invisible null value
}
\description{
Front end to the \code{graphics::segments} function allowing
simplified specification of end points
}
\examples{
nam <- c("mai", "mar", "pin", "plt", "usr", "xaxp", "yaxp", "bg")
oldPar <- par(nam)
par(mar = c(4,4,1,1)+0.1, bg = alpha("sky blue", 0.25))
z <- with(roundTrip, setNames(complex(real = Longitude, imaginary = Latitude), Locality))
plot(z, asp = 1, pch = 20, cex = 0.7, bty = "n", cex.axis = 0.8,
  xlab = "Longitude", ylab = "Latitude", ylim = -c(50,5))
lines(Oz, col = alpha("dark green", 0.5))
text(z, labels = names(z), pos = avoid(z), cex = 0.7)
segments(z, cyc(z), col = "red")
if(lazyData::requireData(PBSmapping)) {
  plotPolys(worldLLhigh, xlim = c(110, 155), ylim = c(-45, -9),
    col = alpha("sandy brown", 0.5), projection = "LL", axes = FALSE,
    border = alpha("navy", 0.25), xlab = "", ylab = "", bty = "n")
  axis(2, at = 10*(-4:-1), cex.axis = 0.7, line = 2)
  axis(1, at = 10*(12:15), cex.axis = 0.7, line = 2)
  points(z, pch = 20, cex = 0.7)
  text(z, labels = names(z), pos = avoid(z), cex = 0.7)
  segments(z, cyc(z), col = "red")
}
  par(oldPar)
}
