% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{showColors}
\alias{showColors}
\alias{showColors.default}
\alias{showColors.function}
\alias{showColours}
\title{Display colours in a collection}
\usage{
showColors(fill, ...)

\method{showColors}{default}(
  fill,
  labels = if (length(fill) <= 25) {     if (!is.null(names(fill)))         
    names(fill)     else fill } else NA,
  border = colour,
  colour = color,
  color = if (length(fill) <= 20) "grey" else "transparent",
  edges = 1000,
  radius = 1,
  ...,
  n = length(fill)
)

\method{showColors}{`function`}(fill, ..., n = 20)

showColours(fill, ...)
}
\arguments{
\item{fill}{either a palette function or a vector of colours to display}

\item{...}{arguments to be passed on to methods}

\item{border, colour, color}{alternative names to use for the colour of the borders}

\item{edges}{integer: no of edges to use for the inner circle}

\item{radius}{numeric: value to use for the radius of the circular display}

\item{n}{number of colours to use for a palette display}
}
\value{
the vector of colours used in the display, invisibly.
}
\description{
Display colours in a collection
}
\examples{
showColors(pal_desert)
pal <- myPalette(pink, beige, 'sky blue')
showColours(pal)
showColours(pal, n = 50)
}
