% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{surfaceColors}
\alias{surfaceColors}
\alias{surfaceColors.list}
\alias{surfaceColors.matrix}
\alias{surfaceColours}
\title{Surface colours for perspective plots}
\usage{
surfaceColors(z, ...)

\method{surfaceColors}{list}(z, ...)

\method{surfaceColors}{matrix}(
  z,
  pal = colorRampPalette(c("steel blue", "beige", "orange red")),
  no = min(250, length(unique(z))),
  col = fill,
  fill = pal(no),
  rgl = FALSE,
  ...
)

surfaceColours(z, ...)
}
\arguments{
\item{z}{matrix or list structure to specify surface}

\item{...}{additional arguments passed on to methods (currently not used)}

\item{pal}{a palette function to use}

\item{no}{the number of colours to use}

\item{col, fill}{alternative names for the vector of colours}

\item{rgl}{logical: should \code{rgl} protocols be used rather than \code{persp}?}
}
\value{
a suitable vector of colours
}
\description{
Generates a suitable colour vector for use with \code{graphics::persp}
}
\examples{
cols <- surfaceColors(volcano, pal = pal_blue2red)
}
