% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{text}
\alias{text}
\alias{text,xy-method}
\alias{text,character-method}
\title{Text}
\usage{
text(x, ...)

\S4method{text}{xy}(x, ...)

\S4method{text}{character}(x, labels, ..., adj, pos, guard = 0.0125)
}
\arguments{
\item{x}{main argument: either numeric, of class \code{"xy"}, or a character string}

\item{...}{}

\item{labels}{text string(s) to be used as labels, ax for \code{graphics::text}}

\item{adj, pos}{as for \code{graphics::text}}

\item{guard}{fraction of the axis length left clear for the character case of \code{x}}
}
\value{
as for \code{graphics::text}
}
\description{
Interface to \code{grpahics::text} allowing more flexibility.  The
character string method allows text to be placed in defined regions
of the plotting surface in a way similar to that allowed by
\code{graphics::legend}
}
\examples{
plot(1:10)
text("top", "(a)", font = 2, family = "serif", cex = 1.5)
text("top left", c("Some text", "to go in", "this corner."), family = "mono")
}
